+++ urlfix/urlfix.py
@@ -40,7 +40,7 @@ def show_parsed_urls(input_file):
 # Keep url, replacement as pair and replace downstream?
 # Directly replace url in document?
 
-def replace_urls(input_file, output_file="replacement.txt", inplace=False, verbose=False, return_matched=False):
+def replace_urls(input_file, output_file="replacement.txt", inplace=False, verbose=False):
     """
 
     :param input_file File whose links you need to replace or check.
@@ -50,9 +50,9 @@ def replace_urls(input_file, output_file="replacement.txt", inplace=False, verbo
 
     :param verbose Logical. Should you be notified of what URLs have moved? Defaults to False.
 
-    :param return_matched Logical. Useful if you want to test on the matched URLs
 
-    :return  File with links replaced.
+    :return  Replaces outdated URL and writes to the specified file. It also returns the number of URLs that have
+    changed. The latter is useful for tests.
 
 
     """
@@ -64,7 +64,7 @@ def replace_urls(input_file, output_file="replacement.txt", inplace=False, verbo
     number_moved = 0
     for index, target_url in enumerate(matched_urls):
 
-        visited = urllib.request.urlopen(Request(target_url, headers={'User-Agent': 'Mozilla'}))
+        visited = urllib.request.urlopen(Request(target_url, headers={'User-Agent': 'XYZ/3.0'}))
         url_used = visited.geturl()
 
         if url_used != target_url:
@@ -73,11 +73,11 @@ def replace_urls(input_file, output_file="replacement.txt", inplace=False, verbo
             with open(input_file, "r") as input_f, open(output_file, "w") as out_f:
                 for line in input_f:
                     if verbose:
-                        print(f"{target_url} replaced with {url_used} in {out_f}")
+                        print(f"{target_url} replaced with {url_used} in {out_f.name}")
                     out_f.write(line.replace(target_url, url_used))
 
             matched_urls[index] = url_used
             number_moved += 1
     information = "URLS have changed" if number_moved != 1 else "URL has changed"
     print(f"{number_moved} {information}")
-    return matched_urls if not return_matched else matched_urls, number_moved
+    return number_moved