+++ docs/source/changelog.rst
@@ -6,6 +6,15 @@ Welcome to urlfix's changelog
 
 
 * 
+  ``returned_matched`` was dropped in ``replace_urls``. Use ``show_parsed_urls`` for low level returns. 
+
+* 
+  ``verbose`` in ``replace_urls`` is now more human friendly by providing the actual name of the output file.
+
+* 
+  Updated ``testurls.md`` to ensure only markdown like links are replaced.
+
+* 
   The regular expression in ``find_links`` was replaced with a more robust one. 
 
 *